0.0.5.4

+++ lib/core/services/api/basket/basket_model.dart
@@ -9,7 +9,7 @@ class FuserResponse {
   @JsonKey(name: 'status')
   int? status;
   @JsonKey(name: 'result')
-  double? result;
+  int? result;
 
   FuserResponse({
     required this.status,
